+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 * feat(ProseSteps): new component
 * feat(ProseCard): new component
 * feat(ProseCardGroup): new component
+* feat(ProseCodeGroup): new component
 * feat(Banner): new component
 * feat(Card): new component
 * feat(CheckboxGroup): new component